+++ test/libsolidity/syntaxTests/parsing/elementary_non_address_state_mutability_state_variable.sol
@@ -1,0 +1,85 @@
+contract C {
+    bool payable a;
+    string payable b;
+    int payable c;
+    int256 payable d;
+    uint payable e;
+    uint256 payable f;
+    bytes1 payable g;
+    bytes payable h;
+    bytes32 payable i;
+    fixed payable j;
+    fixed80x80 payable k;
+    ufixed payable l;
+    ufixed80x80 payable m;
+}
+contract C2 {
+    bool view a;
+    string view b;
+    int view c;
+    int256 view d;
+    uint view e;
+    uint256 view f;
+    bytes1 view g;
+    bytes view h;
+    bytes32 view i;
+    fixed view j;
+    fixed80x80 view k;
+    ufixed view l;
+    ufixed80x80 view m;
+}
+contract C3 {
+    bool pure a;
+    string pure b;
+    int pure c;
+    int256 pure d;
+    uint pure e;
+    uint256 pure f;
+    bytes1 pure g;
+    bytes pure h;
+    bytes32 pure i;
+    fixed pure j;
+    fixed80x80 pure k;
+    ufixed pure l;
+    ufixed80x80 pure m;
+}
+// ----
+// ParserError 9106: (22-29): State mutability can only be specified for address types.
+// ParserError 9106: (44-51): State mutability can only be specified for address types.
+// ParserError 9106: (63-70): State mutability can only be specified for address types.
+// ParserError 9106: (85-92): State mutability can only be specified for address types.
+// ParserError 9106: (105-112): State mutability can only be specified for address types.
+// ParserError 9106: (128-135): State mutability can only be specified for address types.
+// ParserError 9106: (150-157): State mutability can only be specified for address types.
+// ParserError 9106: (171-178): State mutability can only be specified for address types.
+// ParserError 9106: (194-201): State mutability can only be specified for address types.
+// ParserError 9106: (215-222): State mutability can only be specified for address types.
+// ParserError 9106: (241-248): State mutability can only be specified for address types.
+// ParserError 9106: (263-270): State mutability can only be specified for address types.
+// ParserError 9106: (290-297): State mutability can only be specified for address types.
+// ParserError 9106: (326-330): State mutability can only be specified for address types.
+// ParserError 9106: (345-349): State mutability can only be specified for address types.
+// ParserError 9106: (361-365): State mutability can only be specified for address types.
+// ParserError 9106: (380-384): State mutability can only be specified for address types.
+// ParserError 9106: (397-401): State mutability can only be specified for address types.
+// ParserError 9106: (417-421): State mutability can only be specified for address types.
+// ParserError 9106: (436-440): State mutability can only be specified for address types.
+// ParserError 9106: (454-458): State mutability can only be specified for address types.
+// ParserError 9106: (474-478): State mutability can only be specified for address types.
+// ParserError 9106: (492-496): State mutability can only be specified for address types.
+// ParserError 9106: (515-519): State mutability can only be specified for address types.
+// ParserError 9106: (534-538): State mutability can only be specified for address types.
+// ParserError 9106: (558-562): State mutability can only be specified for address types.
+// ParserError 9106: (591-595): State mutability can only be specified for address types.
+// ParserError 9106: (610-614): State mutability can only be specified for address types.
+// ParserError 9106: (626-630): State mutability can only be specified for address types.
+// ParserError 9106: (645-649): State mutability can only be specified for address types.
+// ParserError 9106: (662-666): State mutability can only be specified for address types.
+// ParserError 9106: (682-686): State mutability can only be specified for address types.
+// ParserError 9106: (701-705): State mutability can only be specified for address types.
+// ParserError 9106: (719-723): State mutability can only be specified for address types.
+// ParserError 9106: (739-743): State mutability can only be specified for address types.
+// ParserError 9106: (757-761): State mutability can only be specified for address types.
+// ParserError 9106: (780-784): State mutability can only be specified for address types.
+// ParserError 9106: (799-803): State mutability can only be specified for address types.
+// ParserError 9106: (823-827): State mutability can only be specified for address types.