+++ test/libsolidity/syntaxTests/parsing/payable_accessor.sol
@@ -1,5 +1,0 @@
-contract test {
-	uint payable x;
-}
-// ----
-// ParserError 9106: (22-29): State mutability can only be specified for address types.